pyproject.toml
[tool.poetry]
name = "image-browser"
version = "0.1.0"
description = ""
authors = ["Jason Swift <anemele@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
filetype = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
